Install required ESLint packages
Installed all requireded packages (`devDependencies`):

- babel-eslint
- eslint
- eslint-config-arcticicestudio
- eslint-plugin-babel
- eslint-plugin-import
- eslint-plugin-jsx-a11y
- eslint-plugin-react

>>> Configuration Preset

The eslint-config-arcticicestudio (1) configuration preset implements
the Arctic Ice Studio JavaScript Style (2). It it built on top of
eslint-config-arcticicestudio-base (3) and includes various rules of the
following plugins:

- eslint-plugin-import (4)
- eslint-plugin-jsx-a11y (5)
- eslint-plugin-react (6)

Since the custom preset is still in major version `0` note that the
version range is `>=0.x.x <1.0.0` to avoid NPM's  "SemVer Major Zero
Caveat". When defining package versions with the the carat `^` or tilde
`~` range selector it won't affect packages with a major version of `0`.
NPM will resolve these packages to their exact version until the major
version is greater or equal to `1`.
To avoid this caveat the more detailed version range `>=0.x.x <1.0.0` is
used to resolve all versions greater or equal to `0.x.x` but less than
`1.0.0`. This will always use the latest `0.x.x` version and removes the
need to increment the version manually on each new release.

To allow to lint all valid Babel code babel-eslint (7) will be included
and specified as main parser (8). Also to make use of the latest Babel
features and proposals eslint-plugin-babel (9) has been added.

References:

  (1) https://github.com/arcticicestudio/eslint-config-arcticicestudio
  (2) https://arcticicestudio.github.io/styleguide-javascript
  (3) https://github.com/arcticicestudio/eslint-config-arcticicestudio-base
  (4) https://github.com/benmosher/eslint-plugin-import
  (5) https://github.com/evcohen/eslint-plugin-jsx-a11y
  (6) https://github.com/evcohen/eslint-plugin-react
  (7) https://github.com/babel/babel-eslint
  (8) https://eslint.org/docs/user-guide/configuring#specifying-parser
  (9) https://github.com/babel/eslint-plugin-babel

GH-15