Implement base constants and Gatsby configuration with basic plugins
All metadata and constants, like defined in GH-26, are
placed in the corresponding folders for configurations (`src/config`)
and data (`src/data`). Initially this includes information stored in
the `package.json` file of the project and the `nord` package which has
also been installed.
This commit also adds base and important constants that will be used
later on:

- `src/config/internal/constants.js` - Provides internally used
  constants
- `src/config/internal/nodes.js` - Provides internally used data about
  custom Gatsby GraphQL API nodes.
- `src/config/routes/constants.js` - Provides routing constants.
- `src/config/routes/mappings.js` - Provides route mapping constants.

GH-27