Install React `prop-types` package
For type validation and checking, Nord Docs will for the time being make
use of React prop types (1) through the `prop-types` (2) package. This
technique is sufficient for the beginning of the project and might be
revised and refactored later on e.g. by migrating to TypeScript (3) or
integrating Flow (4). Details will be documented in related future
epics.

References:
  (1) https://reactjs.org/docs/typechecking-with-proptypes.html
  (2) [npm-prop-types]: https://www.npmjs.com/package/prop-types
  (3) https://www.typescriptlang.org
  (4) https://flow.org

Closes GH-35