Implement "Root" core container component
It renders the passed `children` within a `React.Fragment`.

The `Root` core container is the first and one of the important main
data provider React components of this project. It represents the base
element and entry point that wraps the all custom application components
and serves as a data provider later on. Some of the tasks in the future
will be to

- provide context props and functions for the global theme through
  styled-components (1) `ThemeProvider` (3) component
- inject global styles through styled-components
  `createglobalstyle` (3) function including typography (GH-2) related
  data like application-wide used fonts
- possibly data stores for state management libraries like MobX (4)

References:
  (1) https://styled-components.com
  (2) https://www.styled-components.com/docs/api#createglobalstyle
  (3) https://www.styled-components.com/docs/api#themeprovider
  (4) https://mobx.js.org

GH-36