Fix overridden ESLint "import/no-extraneous-dependencies" rule
The import/no-extraneous-dependencies (1) rule allows to define a array
of glob pattern that are allowed to define import `devDependencies`. In
GH-15 the rule was overridden to include the project specific path
`**/.gatsby/**` which whitelists all Gatsby specific scripts.
Unfortunately this removes all glob pattern defined in the used
"eslint-config-arcticicestudio" (2) (rule is defined in the `-base`
package) resulting in errors in other projects paths like tests.
This has been fixed by importing the paths defined in the preset from
the "eslint-config-arcticicestudio-base" (3) package and merge it with
the additional **/.gatsby/**` path.

References:
  (1) https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-extraneous-dependencies.md
  (2) https://www.npmjs.com/package/eslint-config-arcticicestudio
  (3) https://www.npmjs.com/package/eslint-config-arcticicestudio-base

Fixes GH-41