Implement Netlify configuration
This commit implements a part of the implementation of the "hosting &
continuous deployment concept" (1) with Netlify's configuration
file (2). It initially defines the `command` and `publish` keys of the
`[build]` section for production builds.

See the "Hosting" and "Continuous Deployment" (sub)sections of the
related epic issue for more details about the architecture.

References:
  (1) https://github.com/arcticicestudio/nord-docs/issues/46
  (2) https://www.netlify.com/docs/netlify-toml-reference

Related epics: GH-46
Resolves GH-48