Implement basic typography theme properties and values
In order to use the fonts is must be added to Nord Docs theme. This
commit implements the `typography` module that defines all used font
families and the basic typography properties and values like the font
size and units based on the used modular scale documented in GH-2.

Associated epic: GH-2
Dependency of GH-53
GH-54