Use explicit CSS module imports for typefaces
Instead of using the default entry point of the pakages defined in the
`main` field the CSS modules should be imported directly. This fixes the
imports for tests with Jest that are then handled with the implemented
mock for file imports. It improves the DX by showing that these packages
are CSS styles instead of a JS module and removes the need to add
exception/ignore comments for ESLint. The only disadvantage it that it
might be necessary to update the import when the package maintainer
changes the `main` file when updating to a newer version.

GH-53