Implement base HTML element component `A`
The core atom `A` represents the inline text semantic (1) HTML element
`<a>`. It is a special  dynamic and failsafe component since it
internally uses the Gatsby `Link` component to route within the site
(internal links) while also being able to link to external targets. The
URLs are passed through the `to` and `href` props.

References:
  (1) https://developer.mozilla.org/en-US/docs/Web/HTML/Element#Inline_text_semantics
  (2) https://www.gatsbyjs.org/docs/gatsby-link

Associated epic: GH-69
GH-70