Implement the `ErrorState404` component
This commit implements the main 404 (1) page component to handle
requested but non-existing routes and resources. Iz aligns with
Nord Docs brand and can use a humorous illustration and tagline that
conveys the purpose of the screen, without appearing actionable.

See the great Gatsby documentation about 404 pages (2) for more details.

> Structure

The component consists of three parts: an vector illustration, the
headline and the corresponding subline.

> Vector Illustration

The illustration is the main element of the page component. It presents
the exception/error case with a friendly visualization.

> Text Content

The headline and subline inform the user about the purpose of the
illustration and the reason why it is currently shown and can include
helpful instructions on how to continue.

References:
  (1) https://en.wikipedia.org/wiki/HTTP_404
  (2) https://www.gatsbyjs.org/docs/add-404-page

Associated epics: GH-63
GH-84