Implement Gatsby plugins "web app manifest" configueation
Integrated gatsby-plugin-manifest (1) that adds support for generating
and shipping a web app (PWA) `manifest.webmanifest` (2). It allows users
to save Nord Docs as web application to their desktops / task bars and
smartphone home screens so it behaves similar to native apps.

See the great Google Developer documentation about the web app
manifest (3) and the official W3 Web App Manifest specification (4).

Note that is recommended to use this plugin together with
gatsby-plugin-offline (5) (GH-93), but ensure that this plugin is listed
before the offline plugin so that it can cache the created
`manifest.webmanifest` file!

This commit also includes the Nord logo in resized SVG format
(2000x2000) to make sure the generated PNG icons have a high quality.

References:
  (1) https://www.npmjs.com/package/gatsby-plugin-manifest
  (2) https://developer.mozilla.org/en-US/docs/Web/Manifest
  (3) https://developers.google.com/web/fundamentals/web-app-manifest
  (4) https://www.w3.org/TR/appmanifest
  (5) https://www.npmjs.com/package/gatsby-plugin-offline

GH-91