Enable "gatsby-plugin-offline"
Integrated gatsby-plugin-offline (1) that adds drop-in support for
making a Gatsby site work offline and more resistant to bad network
connections. It creates a service worker for the site using the great
Workbox (2) project and loads the service worker into the client.

Note that is recommended to use this plugin together with
gatsby-plugin-manifest (3) (GH-91), but ensure that this plugin is
listed after the offline plugin so that the created
`manifest.webmanifest` file can be included in the service worker!

The plugin comes with optimal configurations out-of-the-box. Note that
the offline support can break when changing these options.

References:
  (1) https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline
  (2)  https://developers.google.com/web/tools/workbox
  (3) https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest
  (4) https://www.npmjs.com/package/gatsby-plugin-offline

GH-94