Implement core atom component to represent a color palette
This is a card that renders a color box with the associated color code
as label and the title of a color palette.
The title can either float above the card (`floatingTitle` prop) or can
be placed within the card below the color code labels using the `title`
prop.

GH-112