Implement main organism components for all landing page sections
>> Hero

The first section is the "hero" of the landing page that renders a short
project description and the animated SVG component
"Arctic Landscape" (1). The animation is triggered through the
"React Waypoint" (1) project.

>> Color Palettes

The 2nd and 3rd sections are about Nord's color palettes and they
provide information about the modularity and contrast. All palettes are
visualized through the new `ColorPaletteCard` component that consists of
the actual color as small, rounded boxes and the associated color code
as label using the hexadecimal format.

>> "Nordify"

The 4th section presents Nord's port projects which can be used to
"nordify" the favorite apps and platforms of the users. It consists of
the text content as well as a new illustration that might also be
animated later on using a 3D parallax effect that renders based on the
current mouse position in the section.

>> Color Swatches

The 5th section is about Nord's "native" color swatches. It consists of
the same elements like the 4th section, but renders a new non-animated
illustration showing a application UI with Nord loaded as color palette.

>> Syntax Highlighting

The 6th section is all about Nord's features to highlight code. Next to
the text content and summary it also renders a new and animated
illustration.

>> Community

The 7th and also last section presents details about Nord's communities.
It renders the animated Arctic Ocean Fractal (3) component and a short
summary about how to join and being part of the Nord community.

>> Responsive Design

All implemented sections follow the responsive design concept (4) to
adjust the rendered content based on the available width and provide an
optimal UX on smaller viewports.

References:
  (1) https://github.com/arcticicestudio/arctic-landscape
  (2) https://github.com/brigade/react-waypoint
  (3) https://www.npmjs.com/package/arctic-ocean-fractal
  (4) https://github.com/arcticicestudio/nord-docs/issues/52

Associated epics: GH-52, GH-63
GH-112