Update to Gatsby 2.1.0 for React Hooks Support
Update to Gatsby v2.1.0 which introduces `useStaticQuery` (1) , a new
Gatsby feature that provides the ability to use a React Hook (2) to
query with GraphQL at build time. It provides the same functionality
like the `StaticQuery` (3) component, but as a Hook rather than a
component that takes a render prop.

References:
  (1) https://www.gatsbyjs.org/docs/use-static-query
  (2) https://reactjs.org/docs/hooks-intro.html
  (3) https://www.gatsbyjs.org/docs/static-query

Resolves GH-125