Implemented interactive and animated port project SVG illustration
A static vector graphic that is currently placed in the port section of
the landing (`/`) page has been implemented as interactive animated
component. It uses a 3D depth effect for the port project's logos by
moving against the direction of the current mouse position where the
larger logos will move more to give the impression that they are closer.

For the interactive animation the awesome react-spring (1) project is
used which allows, like the fantastic and also already used react-pose
(2) library, to use spring-physics based animations.
The advantage of react-spring is that it bails out of the React renderer
to avoid unnecessary renders and instead uses native methods to get the
best performance even for complex animations.
It comes with first-class support for Hooks (3), a great and extensive
documentation and a lot of examples for inspirations (4).

References:
  (1) https://www.react-spring.io
  (2) https://popmotion.io/pose
  (3) https://www.react-spring.io/docs/hooks/basics
  (4) https://www.react-spring.io/docs/hooks/examples

GH-119