Implement hero and card grid filter section components
The "Hero" section contains a feature duo container with the pages
heading and a summary about Nord's ports.
Underneath the static vector graphic that is currently placed in the
port section of the landing (`/`) page has been implemented as
interactive animated component. It uses a 3D depth effect for the port
project's logos by moving against the direction of the current mouse
position where the larger logos will move more to give the impression
that they are closer.

The "Grid Filter" section consists of a input field that allows users to
filter available port projects by predefined search terms. Underneath a
animated grid of card components reacts to the input field by filtering
non-matching ports projects.
Each card represents one port project and consists of the logo, the name
(with the "Nord" prefix) as well as a card footer that contains the
version of the latest release of the respective project and the number
of stars of the GitHub repository.

GH-119