Use interactive illustration for ports section of landing page
GH-126 improved the already existing "Nordify" SVG illustration by
making it an interactive components that animates the floating logos in
the opposite direction of the current mouse cursor's position.

This component should now also be used for the [ports section of the
landing page (1) replacing the current static SVG illustration/image.

References:
  (1) https://develop.nordtheme.com/#ports

Related to GH-126
Resolves GH-127