Implement blog post grid and card components, GraphQL nodes and queries
The last remaining main page is `/blog` which presents an overview of
Nord's blog posts in a three-column grid of card components sorted
descending by their publish date. Each card consists of an cover image
together with the title of the post. The latest post spans over two
columns at the top for a nice visual structure and better
recognizability.
While a one-column card uses a cover image the latest post will use a
"banner" that will either be the same image with a larger width,
a variant of it or a completly different one.

A blog post itself makes use of the MDX features and the custom MDX
components mentioned in the paragraph above. To simplify the usage of
the "cover" and "banner" image they are be processed with Gatsby's
`onCreateNode` (1) API in combination with Gatsby's `mapping`
configuration feature (2). This allows to map the paths to the images
to a `File` node that will then be handled by the Gatsby image
processing plugin workflow (3) also documented in the commit about
image handling.

Another required node is the `heroImage` field that queries for a
`hero.(png|jpg|jpeg)` image that is used as the hero of a blog post.
To allow to also use videos in blog posts or even as header a custom
`Video` MDX component has been implemented including the optional
`heroVideo` and `heroVideoPoster` GraphQL node fields.

All together that results in the following required and optional
images/videos mapped to specific node fields with reserved file names
per blog post directory for simple usage via GraphQL queries:

- `bannerImage` <-> `banner.(png|jpg|jpeg)` - The required banner image
  of a blog post card (used when currently the latest two-column wide
  post placed on top of the grid).
- `coverImage` <-> `cover.(png|jpg|jpeg)` - The required cover image of
  a one-column blog post.
- `heroImage` <-> `hero.(png|jpg|jpeg)` - The required hero image of a
  blog post.
- `heroVideo` <-> `hero.(mp4|webm)` - The optional hero video of a blog
  post.
- `heroVideoPoster` <-> `heroposter.(png|jpg|jpeg)` - The optional
  poster image of a blog post `heroVideo`.

>>> Known Problems

To prevent the `unknwon field` GraphQL error during build time (4)
(e.g. when there are no blog posts yet) a "dummy"/"placeholder" blog
post and docs page will be created.
Anyway, this will be removed as soon as there is finally the first blog
post and docs page. Later on the project will migrate to the shiny new
schema customization API (5).

References:
  (1) https://www.gatsbyjs.org/docs/node-apis/#onCreateNode
  (2) https://www.gatsbyjs.org/docs/gatsby-config/#mapping-node-types
  (3) https://www.gatsbyjs.org/docs/working-with-images
  (4) https://github.com/gatsbyjs/gatsby/issues/3344
  (5) https://www.gatsbyjs.org/blog/2019-03-04-new-schema-customization

GH-129